mysite/mysite/requirements.txt
aiohttp==3.7.4
APScheduler==3.8.1
asgiref==3.4.1
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
awscli==1.22.26
awsebcli==3.20.2
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bokeh==2.1.1
botocore==1.23.26
Brotli==1.0.7
bs4==0.0.1
cachelib==0.4.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
dash==1.13.4
dateparser==1.1.0
DateTime==4.3
decorator==4.4.2
discord==1.0.1
discord.py==1.5.1
distlib==0.3.4
Django==4.0
django-apscheduler==0.6.0
docutils==0.15.2
dtale==1.12.0
filelock==3.4.2
flasgger==0.9.4
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.5.0
flask-ngrok==0.0.25
Flask-Session==0.4.0
future==0.16.0
gevent==20.6.2
greenlet==0.4.16
idna==2.10
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
iso8601==1.0.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
keyboard==0.13.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.4
lz4==3.1.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
multidict==4.7.6
multitasking==0.0.9
mysql-connector==2.2.9
mysql-connector-python==8.0.26
netifaces==0.11.0
numpy==1.19.1
packaging==20.4
pandas==1.0.5
parso==0.7.1
pathspec==0.5.9
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.4.1
plotly==4.9.0
prompt-toolkit==3.0.5
protobuf==3.18.1
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pylivestream==1.11.4
pyodbc==4.0.30
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.16.0
python-binance==1.0.15
python-dateutil==2.8.1
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==228
PyYAML==5.4.1
pyzmq==19.0.1
regex==2021.11.10
requests==2.26.0
retrying==1.3.3
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.23.2
scipy==1.5.2
scripts==2.0
selenium==3.141.0
semantic-version==2.8.5
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.2
statsmodels==0.11.1
stua==0.2
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
twitchio==2.1.3
typing-extensions==3.7.4.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
virtualenv==20.13.0
wcwidth==0.1.9
webdriver-manager==3.2.1
websocket==0.2.1
websocket-client==1.2.1
websockets==9.1
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.16.0
yarl==1.5.1
yfinance==0.1.62